Don't install keyring to apt trusted directory
None of the keys in our keyring are used for signing our debian
repository, so it's inappropriate to tell apt to trust repositories
signed with those keys. Any repository signed with one of the keys in
our keyring would have to be malicious since we'd never sign with one of
these keys.