Split up keyring into service specific keyrings
Having a single keyring that contains all of our keys and then using it
for verification in ostree and the app-manager is wrong. For instance,
if an ostree commit was signed directly by one of our master keys or one
of the app updater keys, that would indicate malicious activity.

Furthermore, the purpose of the keyrings is not to contain the entire
chain of trust, only the specific signatures that should be trusted by
the corresponding service. For that reason, the master keys are only
installed to the general keyrings directory where a user can establish
the chain of trust if they desire. This is how debian and ubuntu handle
their keyrings.