Add keyring for codecs activation
Add another keyring, eos-codecs-keyring.gpg, containing the public
part of the keys used for encrypting codecs that will be shipped in
a deactivated form along with our OSTree images.

At the moment this containes the ECAK1 key, which will be the one that
end users will normally use for decryption, as well as the EAMK1 key,
which will only be used by us as a master key.

The keyring is installed only in the generic keyrings directory
since there's no automated image verification process.

[endlessm/eos-shell#5767]