Add endless-ca-cert package with Endless CA root cert
Provide the Endless SSL CA root certificate in the endless-ca-cert
package. The certificate will be added to the system's trusted CA
certificates by triggering update-ca-certificates --fresh. The setup
here is based on the ca-cacert package. The --fresh option is needed so
that the symlinks in /etc/ssl/certs are updated if the certificate
changes or removed when the package is uninstalled.

If endless/endless-ca.crt was previously installed and removed (it was
temporarily in the main ca-certificates package), then "dpkg-reconfigure
ca-certificates" may need to be run to re-enable it.

https://phabricator.endlessm.com/T18743