Fix build with gpg2
A GPG "keyring" is simply a stream of OpenPGP packets. In GnuPG 2.1, gpg
started using a "keybox" to store keys, which is an indexed structure.
Since we import keys directly to the final keyring, it's created as a
"keybox" when gpg2 is in use. This is intended to be an internal
representation and is not the keyring format clients expect. Instead,
import to a temporary keyring and then export to the final keyring. This
produces the expected keyring format since --export produces the proper
OpenPGP packets.

This was broken when gpg became gpg2 on master. See
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=790665#10 for some
further details where apt was dealing with the same issue.

https://phabricator.endlessm.com/T18743