Add script for importing Endless CA cert to user's NSS DB
Chrome and chromium use their own certificate stores with user
customizations stored in ~/.pki/nssdb, so adding to
/etc/ssl/certs/ca-certificates.crt will have no effect for those
browsers.

This script looks for the installed endless-ca.crt file and uses
certutil to import it to ~/.pki/nssdb as a trusted CA. Each user will
have to run this after the cert is downloaded with the
endless-ca-download script.

https://phabricator.endlessm.com/T18743