Provide profile script to run endless-ca-trust-user on login
Installing the endless-ca-cert package or running
endless-ca-trust-system will add the Endless CA certificate to the
system CA certificate bundle, but it won't affect chromium or chrome
that use NSS. The endless-ca-trust-user script can handle that, so
provide an /etc/profile.d script that runs it on login for non-system
users. This will be included in the endless-ca-cert package, so it will
only run when the system is already trusting the Endless CA certificate.

https://phabricator.endlessm.com/T18743