Update flatpak signing key
This updates the flatpak (EFSK1) public key used for signing flatpak
apps and runtimes. The change is to extend the expiration date by 5
years to 2026. That conforms with a new policy that all signing keys
will expire at least 5 years from now.

The updated UID has been signed with the Endless master keys as before.

https://phabricator.endlessm.com/T27034