Add test for key expiration dates
This checks that each key in a generated keyring conforms to our policy
that they will not expire within 5 years. We currently have some keys
that do not meet this policy, so the failures are converted to expected
failures (XFAIL) by adding a TODO comment. Later when we have all keys
conforming, this should be changed back to a hard failure.

https://phabricator.endlessm.com/T27113
(cherry picked from commit 68779b7b38e145557a613fbf54076f06f822931e)