Add .gitattributes file for key diff text conversion
Rather than looking at the diff of a string of base64 characters, set
a diff text conversion program for .pub GPG keys and .crt X.509
certificates. These need to be defined in the git configuration either
in the repo or globally. If the definitions don't exist git just uses
the standard diff algorithm, and to see the raw diff, pass --no-textconv
to git diff.

https://phabricator.endlessm.com/T28413