Add secure boot archive key for testing
The ESBASK1 signing key is used for the internal secure boot signer
archive. This archive holds signed versions of packages that are then
pulled into the main OBS archive.

This key is only used by the internal infrastructure, so it gets built
as part of an uninstalled keyring. This allows using the key expiration
framework without actually adding the key to the OS.

https://phabricator.endlessm.com/T28413