test-x509-expiration: Use x509 -checkend to verify minimum expiration
On 32 bit ARM, using `date` to convert the certificate's expiration date
string to seconds since the epoch fails if the date is beyond 2039.
Instead, use the x509 -checkend option, which takes a minimum number of
seconds until expiration and exits with a failure if the certificate's
expiration is within that time.

https://phabricator.endlessm.com/T28912