efi: Add new EOS Driver Signing 2020 certificate
This will be used to sign out of tree kernel modules rather than the
certificate used to sign kernels and bootloaders. The reason is that
this certificate is embedded in the kernel and can be used to execute
new kernels via the `kexec` mechanism. That would allow the kernel on
PAYG systems to execute EOS kernels and break out of the PAYG security
domain.

https://phabricator.endlessm.com/T30607