Update SSL root CA
It was going to begin violating our 5 year expiration policy in one
week. Now it has a 20 year expiration window. While here, the name
constraints were adjusted so that it cannot issue endlessm.com or
endlessos.com certificates. We no longer need to use this CA to issue
certificates for our public domains.

https://phabricator.endlessm.com/T33545