Add GlobalSign EV code signing certificate
This certificate is only really used for submissions to Microsoft, but
storing it here gives it a well known home. The CSR and GlobalSign CA
certificates forming the chain are also included for reference.

More importantly, it allows testing the expiration date so we don't
forget to renew it. The certificate is only valid for 1 year, so a
special case is needed to lower the expiration window to 30 days.

https://phabricator.endlessm.com/T33943